requirements.txt
pwinput==1.0.3
pycryptodome==3.20.0
